+++ src/main/java/org/magma/name/IntNameResolver.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.inject.Inject;
 import org.magma.Compiler;
-import org.magma.parse.JSONUnit;
+import org.magma.JSONUnit;
 
 import java.util.Optional;
