+++ src/main/java/org/magma/parse/FunctionParser.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import org.magma.Compiler;
+import org.magma.JSONUnit;
 import org.magma.util.Scope;
 
 import java.util.Optional;
